test: set appropriate mapping template version on tests
For tests that check the full resulting object, including the mapping
template version, we set the appropriate API's mapping template version
before it runs.

We do this because we have some tests that will set V2. By doing so,
every test that runs after them will be V2, even the ones that needs to
be V1 (default).

Signed-off-by: Juliana Oliveira <juliana@uma.ni>